Issue-439: Person module created
