+++ app/build.gradle
@@ -54,8 +54,6 @@ dependencies {
     implementation libs.kotlin
     implementation libs.coroutines
 
-    implementation libs.pagerIndicator
-
     implementation libs.bundles.architectureComponents
 
     implementation libs.bundles.di
@@ -74,6 +72,7 @@ dependencies {
     implementation project(path: ':networkdata')
     implementation project(path: ':screen:review')
     implementation project(path: ':screen:country')
+    implementation project(path: ':screen:person')
     implementation project(path: ':view:voteview')
     implementation project(path: ':view:keywordsview')
     implementation project(path: ':view:creditsview')
