--- app/src/main/java/com/kirchhoff/movies/data/ui/details/person/UIPersonImage.kt
@@ -1,3 +1,0 @@
-package com.kirchhoff.movies.data.ui.details.person
-
-data class UIPersonImage(val url: String)
