+++ app/src/main/java/com/kirchhoff/movies/di/RepositoryModule.kt
@@ -4,10 +4,10 @@ import com.kirchhoff.movies.repository.discover.DiscoverRepository
 import com.kirchhoff.movies.repository.discover.IDiscoverRepository
 import com.kirchhoff.movies.repository.movie.IMovieRepository
 import com.kirchhoff.movies.repository.movie.MovieRepository
-import com.kirchhoff.movies.repository.person.IPersonsRepository
-import com.kirchhoff.movies.repository.person.PersonsRepository
 import com.kirchhoff.movies.repository.tv.ITvRepository
 import com.kirchhoff.movies.repository.tv.TvRepository
+import com.kirchhoff.movies.screen.person.repository.IPersonsRepository
+import com.kirchhoff.movies.screen.person.repository.PersonsRepository
 import org.koin.dsl.module
 
 val repositoryModule = module {
