+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/MainPagerAdapter.kt
@@ -3,8 +3,8 @@ package com.kirchhoff.movies.ui.screens.main
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
+import com.kirchhoff.movies.screen.person.ui.screen.list.PersonsFragment
 import com.kirchhoff.movies.ui.screens.main.movies.MoviesFragment
-import com.kirchhoff.movies.ui.screens.main.persons.PersonsFragment
 import com.kirchhoff.movies.ui.screens.main.tvs.TvsFragment
 
 class MainPagerAdapter(fm: FragmentManager) :
