+++ screen/person/src/main/java/com/kirchhoff/movies/screen/person/data/UIPersonImage.kt
@@ -1,0 +1,3 @@
+package com.kirchhoff.movies.screen.person.data
+
+data class UIPersonImage(val url: String)
