+++ screen/person/src/main/java/com/kirchhoff/movies/screen/person/mapper/main/IPersonsMapper.kt
@@ -1,0 +1,11 @@
+package com.kirchhoff.movies.screen.person.mapper.main
+
+import com.kirchhoff.movies.core.data.UIPerson
+import com.kirchhoff.movies.core.repository.Result
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
+import com.kirchhoff.movies.networkdata.core.NetworkPaginated
+import com.kirchhoff.movies.networkdata.main.NetworkPerson
+
+interface IPersonsMapper {
+    fun createUIPersons(persons: Result<NetworkPaginated<NetworkPerson>>): Result<UIPaginated<UIPerson>>
+}
