+++ screen/person/src/main/res/layout/fragment_person_image.xml
