+++ screen/person/src/main/res/layout/item_person.xml
